+++ extensions/qq-mail/CHANGELOG.md
@@ -1,3 +1,9 @@
 # QQ Mail Changelog
 
 ## [Initial Release] - {PR_MERGE_DATE}
+
+- Email browsing
+- Email composing
+- Folder filtering
+- Message actions
+- Attachment downloads